algorithm: Add algo to find the page number of a financial statement.
We want to automate everything eventually or at least minimize
human intervention as much as possible.

Added option to not specify "start" and "end" and rather specify
"statement_name" to automatically find the page number of
the specified statement.

The function matches the availability of most probable keywords
that are found on a particular financial statement.

Only available for "balance sheet" right now.